Test: Fix bip148-segwit-uasf.py
Github-Pull: #10532
Rebased-From: 349f5e92961d057c06cca29538ef270a521995eb
