+++ qa/pull-tester/rpc-tests.py
@@ -103,6 +103,7 @@ testScripts = [
     'wallet-hd.py',
     'walletbackup.py',
     # vv Tests less than 5m vv
+    'bip148-segwit-uasf.py',
     'p2p-fullblocktest.py',
     'fundrawtransaction.py',
     'p2p-compactblocks.py',
